add demo using dynamic UI
This PR adds a new frontend demo, installable via `npx create-reactor-app`. With this demo it's possible
to automatically generate a UI for any Reactor Model ran LOCALLY. It makes use of the new `ReactorController` component, which automatically requires from the model the capabalities, which the model sends back as a JSON schema.
which was added in version 1.0.17 of the JS SDK.

This is especially useful for researchers development, as it allows to quickly test and debug a model without needing to code a full frontend demo.

topic: add-dynamic-ui-demo-example-frontend
reviewers: snow, bryce